Refactor configuration clauses
Rules may now be configured with a list of clauses, both when registering a rule
via the attribute and when adding a rule as part of an Aesop call. Clauses
implemented so far:

- builder <b>: Use the specified rule builder instead of the default.
- safe: Mark the rule as safe. (Only applies to rules in the safe phase.)
- almost_safe: Mark the rule as almost safe. (Only applies to rules in the safe phase.)

Tactic configuration clauses (those given as part of an Aesop call) now have a
different syntax, for consistency with rule clauses and because the previous
syntax messed up the parsing of begin-end blocks.